+++ History.md
@@ -1,5 +1,29 @@
 
-0.0.2 / 2012-07-05 
+0.1.0 / 2012-12-03 
 ==================
 
-  * fix dialog.effect support
+  * add absolute positioning support via .show(x, y)
+  * add Tip#cancelHideOnHover()
+
+0.0.5 / 2012-08-31 
+==================
+
+  * fix hiding of tip when hover back over the target
+
+0.0.4 / 2012-08-22 
+==================
+
+  * fix unnecessary applying of content on .show() [guille]
+
+0.0.3 / 2012-08-22 
+==================
+
+  * add `Tip#attach(el, [delay])`
+  * add `.value` option
+  * change `Tip#cancelHideOnHover()` to be public
+  * fix npm template.js usage
+
+0.0.2 / 2012-08-22 
+==================
+
+  * add `Tip#cancelHideOnHover()`